refactor: simplify AI store and centralize types
- Removed currentMode, switchToAI, and switchToSearch from AIAssistantStore
- Moved DocumentationContextItem, CategoryItem, and AIDocumentationContext to @/types/ai.ts
- Updated all hooks and components to use centralized types
- Cleaned up store tests to remove obsolete mode switching tests
- Verified permanent 301 redirects in next.config.mjs
